user actions and reducer refactored
