+++ server.js
@@ -51,17 +51,16 @@ app.use((req, res, next) => {
   const token = req.session.token || req.body.token || req.headers['x-access-token']
   if (token) {
     jwt.verify(token, app.get('superSecret'), (err, decoded) => {
-        if(err) {
-          return res.json({
-            success: false,
-            message: 'Failed to authenticate token'
-          })
-        } else {
-          req.decoded = decoded
-          next()
-        }
+      if(err) {
+        return res.json({
+          success: false,
+          message: 'Failed to authenticate token'
+        })
+      } else {
+        req.decoded = decoded
+        next()
       }
-    )
+    })
   } else {
     if (req.url === '/login') {
       next()
