+++ src/actions/types.js
@@ -1,9 +1,14 @@
+// user actions
 export const AUTH_USER = 'auth_user'
 export const GET_ACCOUNTS = 'get_accounts'
 export const LOGIN = 'login_user'
 export const VERIFY_SIGNATURE = 'verify_signature'
 export const GET_USER_INFO = 'get_user_info'
 
+
+
+
+// assets actions
 export const GET_EXCHANGE_INFO = 'get_exchange_info'
 export const ADD_NEW_EXCHANGE = 'add_new_exchange'
 export const SAVE_USER_ADDRESS = 'save_user_address'
