+++ src/components/welcome/manager/Welcome.js
@@ -1,13 +1,12 @@
 import React, { Component } from 'react'
 import { connect } from 'react-redux'
-import {AuthActions} from '../../../actions'
-import Presentation from '../view/Welcome'
+import View from '../view/Welcome'
 
 
 class WelcomeScreen extends Component {
     render() {
         return (
-          <Presentation />
+          <View />
         )
     }
 }
