+++ src/reducers/index.js
@@ -1,10 +1,10 @@
-import authReducer from './authReducer'
+import userReducer from './userReducer'
 import exchangeReducer from './exchangeReducer'
 import coinReducer from './coinReducer'
 import walletReducer from './walletReducer'
 import assetReducer from './assetReducer'
 export {
-  authReducer,
+  userReducer,
   exchangeReducer,
   coinReducer,
   walletReducer,
